Completely rework the block parser so it finishes an element before parsing its children.
The parser used to completely parse each line before moving on to the next. But
those poses a problem for the detection of underline-style headers. For
example:

    > The body of a paragraph.
    > A Header
    > --------

The paragraph needs to detect the line of dashes, infer that there's a header,
and end the paragraph early. But it can't if the parser moves line-by-line
because it can't know that it needs to skip the `> ` from the blockquote. This
change will make it possible to correctly handle this case.

Additionally, this change cleans up and improves the readability of the code.