Rewrite MMSpanScanner so that it's a recursive descent parser.
This improves backtracking, which turns out to be an important feature for
correctly handling certain types of input. For instance,

    ***Strong** and Em*

should be transformed into this:

    <em><strong>Strong</strong> and Em</em>

The 2 leading stars would normally indicate a strong, so the parser needs to
backtrack to determine that the leading star should be taken on its own.

Of course, a number of existing parsers get this wrong. But now MMMarkdown
gets it right!

The code is also now cleaner and simpler.