Only backslash escape specific characters.
The markdown specification specifices that backslash escapes are
"provided" for certain characters. The expected behavior is that any
other characters that are preceded by a backslash will be passed
through untouched along with the backslash.

MMMarkdown was always removing the backslash and passing the character
through. Change the code so that the backslash is only removed if the
next character is one of the special characters.