Begin to actually parse HTML
Introduce a new class, MMTHMLParser, to handle the parsing. It will
only handle a subset of valid HTML, but it's enough to begin passing
some failing tests.

MMSpanParser now tries to use this class to parse inline HTML. By
doing so, it will correctly handle attributes which look like they
contain markdown, such as this:

    <a href="#" title="*Bold Title*">A Link</a>

In such a case, the asterisks should be passed through unchanged.