Fix an incorrect HTML test by doing a string comparison
The test was written incorrectly before. After correcting the test by
dropping the double-quotes, it was no longer valid XML. So the existing
test function, MMAssertMarkdownEqualsHTML, would pass because both the
generated HTML and the expected HTML would fail to parse and return
nil.

Introduce a new test function, MMAssertMarkdownEqualsString, which
performs a string comparison on the actual and expected outputs.